+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # 版本变更说明
 
+## v1.1.9 - 2022.4.19
+
+- fix(FormControl): 修复表单校验错误为`object`类型时，页面空白的 bug
+
+## v1.1.8 - 2022.4.8
+
+- improve(Scrollbar): 出现滚动条时，分别添加对应的 class 名称
+
+## v1.1.7 - 2022.4.1
+
+- feat(TextInput): TextInput 在 outline 模式下支持方角显示
+
 ## v1.1.6 - 2022.3.17
 
 - fix(Scrollbar): 修复子元素发生变化，没有自动计算滚动条布局的缺陷